add code
добавлена обработка прайсов РИВАЛЬ (подкрылки и подлокотники). Небольшой
рефакторинг обработчика прайсов РИВАЛЬ.
