+++ UI/EnumPrices.cs
@@ -11,6 +11,8 @@
         Неизвестный = 6,
         Левандовская = 7,
         Автовентури = 8,
-        АвтоБРОНЯ = 9
+        РИВАЛЬ_АвтоБРОНЯ = 9,
+        РИВАЛЬ_Подкрылки = 10,
+        РИВАЛЬ_Подлокотники = 11
     }
 }
