+++ UI/ExcelParser.cs
@@ -155,11 +155,24 @@ namespace ExcelParserForOpenCart
                     _resultingPrice = PTGrupp.ResultingPrice;
                     break;
 
-                case EnumPrices.АвтоБРОНЯ:
-                    var AutoBRONYA = new AutoBronya(sender, e);
-                    AutoBRONYA.Analyze(row, range);
+                case EnumPrices.РИВАЛЬ_АвтоБРОНЯ:
+                    var AutoBRONYA = new Rival(sender, e);
+                    AutoBRONYA.AnalyzeBronya(row, range);
                     _resultingPrice = AutoBRONYA.ResultingPrice;
                     break;
+
+                case EnumPrices.РИВАЛЬ_Подкрылки:
+                    var Podkrilki = new Rival(sender, e);
+                    Podkrilki.AnalyzePodkrilki(row, range);
+                    _resultingPrice = Podkrilki.ResultingPrice;
+                    break;
+
+                case EnumPrices.РИВАЛЬ_Подлокотники:
+                    var Podlokotniki = new Rival(sender, e);
+                    Podlokotniki.AnalyzePodlokotniki(row, range);
+                    _resultingPrice = Podlokotniki.ResultingPrice;
+                    break;
+
 
                 case EnumPrices.Autogur73:
                     var autogurPrice = new AutogurPrice(sender, e);
