+++ UI/ExcelParserForOpenCart.csproj
@@ -105,7 +105,7 @@
   <ItemGroup>
     <Compile Include="OutputPriceLine.cs" />
     <Compile Include="PairProductAndCost.cs" />
-    <Compile Include="Prices\AutoBronya.cs" />
+    <Compile Include="Prices\Rival.cs" />
     <Compile Include="Prices\AutogurPrice.cs" />
     <Compile Include="Prices\Autoventuri.cs" />
     <Compile Include="Prices\For2Union.cs" />
